Reorganize examples and lib files
