+++ component_template/src/streamlit/index.tsx
@@ -1,6 +1,6 @@
 /**
  * @license
- * Copyright 2018-2019 Streamlit Inc.
+ * Copyright 2018-2020 Streamlit Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -15,4 +15,10 @@
  * limitations under the License.
  */
 
-export { default } from "./Table"
+import ArrowTable from "./ArrowTable"
+import { ComponentProps } from "./StreamlitComponent"
+
+export { StreamlitComponent } from "./StreamlitComponent"
+export { connectToStreamlit } from "./connectToStreamlit"
+export type ArrowTable = ArrowTable
+export type ComponentProps = ComponentProps
